refactor(amazonq): move tool execution logic to a separate class
