+++ server/aws-lsp-codewhisperer/src/language-server/agenticChat/agenticChatController.test.ts
@@ -604,517 +604,6 @@ describe('AgenticChatController', () => {
             sinon.assert.notCalled(addMessageStub)
         })
 
-        it('handles tool use responses and makes multiple requests', async () => {
-            // First response includes a tool use request
-            const mockToolUseId = 'mock-tool-use-id'
-            const mockToolName = 'mock-tool-name'
-            const mockToolInput = JSON.stringify({ param1: 'value1' })
-            const mockToolResult = { result: 'tool execution result' }
-
-            const mockToolUseResponseList: ChatResponseStream[] = [
-                {
-                    messageMetadataEvent: {
-                        conversationId: mockConversationId,
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'I need to use a tool. ',
-                    },
-                },
-                {
-                    toolUseEvent: {
-                        toolUseId: mockToolUseId,
-                        name: mockToolName,
-                        input: mockToolInput,
-                        stop: true,
-                    },
-                },
-            ]
-
-            // Second response after tool execution
-            const mockFinalResponseList: ChatResponseStream[] = [
-                {
-                    messageMetadataEvent: {
-                        conversationId: mockConversationId,
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'Hello ',
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'World',
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: '!',
-                    },
-                },
-            ]
-
-            chatDbInitializedStub.returns(true)
-            getMessagesStub.onFirstCall().returns([])
-            getMessagesStub.onSecondCall().returns([
-                {
-                    type: 'prompt',
-                    body: 'Hello with tool',
-                    userInputMessageContext: {
-                        toolResults: [],
-                    },
-                },
-                {
-                    type: 'answer',
-                    body: 'I need to use a tool. ',
-                    toolUses: [{ toolUseId: mockToolUseId, name: mockToolName, input: { key: mockToolInput } }],
-                },
-            ])
-
-            // Reset the stub and set up to return different responses on consecutive calls
-            generateAssistantResponseStub.restore()
-            generateAssistantResponseStub = sinon.stub(CodeWhispererStreaming.prototype, 'generateAssistantResponse')
-
-            generateAssistantResponseStub.onFirstCall().returns(
-                Promise.resolve({
-                    $metadata: {
-                        requestId: mockMessageId,
-                    },
-                    generateAssistantResponseResponse: createIterableResponse(mockToolUseResponseList),
-                })
-            )
-
-            generateAssistantResponseStub.onSecondCall().returns(
-                Promise.resolve({
-                    $metadata: {
-                        requestId: mockMessageId,
-                    },
-                    generateAssistantResponseResponse: createIterableResponse(mockFinalResponseList),
-                })
-            )
-
-            // Reset the runTool stub
-            const runToolStub = testFeatures.agent.runTool as sinon.SinonStub
-            runToolStub.reset()
-            runToolStub.resolves(mockToolResult)
-
-            // Make the request
-            const chatResultPromise = chatController.onChatPrompt(
-                { tabId: mockTabId, prompt: { prompt: 'Hello with tool' } },
-                mockCancellationToken
-            )
-
-            const chatResult = await chatResultPromise
-
-            // Verify that generateAssistantResponse was called twice
-            sinon.assert.calledTwice(generateAssistantResponseStub)
-
-            // Verify that the second request included the tool results in the userInputMessageContext
-            const secondCallArgs = generateAssistantResponseStub.secondCall.args[0]
-            assert.ok(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-                    .length,
-                1
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].toolUseId,
-                mockToolUseId
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].status,
-                'success'
-            )
-            assert.deepStrictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].content[0].json,
-                mockToolResult
-            )
-
-            // Verify that the history was updated correctly
-            assert.ok(secondCallArgs.conversationState?.history)
-            assert.strictEqual(secondCallArgs.conversationState?.history.length, 2)
-            assert.ok(secondCallArgs.conversationState?.history[0].userInputMessage)
-            assert.ok(secondCallArgs.conversationState?.history[1].assistantResponseMessage)
-
-            // Verify the final result
-            assertChatResultsMatch(chatResult, expectedCompleteChatResult)
-        })
-
-        it('propagates tool execution errors to the model in toolResults', async () => {
-            // First response includes a tool use request
-            const mockToolUseId = 'mock-tool-use-id-error'
-            const mockToolName = 'mock-tool-name'
-            const mockToolInput = JSON.stringify({ param1: 'value1' })
-            const mockErrorMessage = 'Tool execution failed with an error'
-
-            const mockToolUseResponseList: ChatResponseStream[] = [
-                {
-                    messageMetadataEvent: {
-                        conversationId: mockConversationId,
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'I need to use a tool that will fail. ',
-                    },
-                },
-                {
-                    toolUseEvent: {
-                        toolUseId: mockToolUseId,
-                        name: mockToolName,
-                        input: mockToolInput,
-                        stop: true,
-                    },
-                },
-            ]
-
-            // Second response after tool execution error
-            const mockFinalResponseList: ChatResponseStream[] = [
-                {
-                    messageMetadataEvent: {
-                        conversationId: mockConversationId,
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'I see the tool failed with error: ',
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: mockErrorMessage,
-                    },
-                },
-            ]
-
-            chatDbInitializedStub.returns(true)
-            getMessagesStub
-                .onFirstCall()
-                .returns([])
-                .onSecondCall()
-                .returns([
-                    {
-                        type: 'prompt',
-                        body: 'Hello with failing tool',
-                        userInputMessageContext: {
-                            toolResults: [],
-                        },
-                    },
-                    {
-                        type: 'answer',
-                        body: 'I need to use a tool that will fail. ',
-                        toolUses: [{ toolUseId: mockToolUseId, name: mockToolName, input: { key: mockToolInput } }],
-                    },
-                ])
-
-            // Reset the stub and set up to return different responses on consecutive calls
-            generateAssistantResponseStub.restore()
-            generateAssistantResponseStub = sinon.stub(CodeWhispererStreaming.prototype, 'generateAssistantResponse')
-
-            generateAssistantResponseStub.onFirstCall().returns(
-                Promise.resolve({
-                    $metadata: {
-                        requestId: mockMessageId,
-                    },
-                    generateAssistantResponseResponse: createIterableResponse(mockToolUseResponseList),
-                })
-            )
-
-            generateAssistantResponseStub.onSecondCall().returns(
-                Promise.resolve({
-                    $metadata: {
-                        requestId: mockMessageId,
-                    },
-                    generateAssistantResponseResponse: createIterableResponse(mockFinalResponseList),
-                })
-            )
-
-            // Reset the runTool stub and make it throw an error
-            const runToolStub = testFeatures.agent.runTool as sinon.SinonStub
-            runToolStub.reset()
-            runToolStub.rejects(new Error(mockErrorMessage))
-
-            // Make the request
-            const chatResultPromise = chatController.onChatPrompt(
-                { tabId: mockTabId, prompt: { prompt: 'Hello with failing tool' } },
-                mockCancellationToken
-            )
-
-            const chatResult = await chatResultPromise
-
-            // Verify that generateAssistantResponse was called twice
-            sinon.assert.calledTwice(generateAssistantResponseStub)
-
-            // Verify that the second request included the tool error in the toolResults with status 'error'
-            const secondCallArgs = generateAssistantResponseStub.secondCall.args[0]
-            assert.ok(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-                    .length,
-                1
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].toolUseId,
-                mockToolUseId
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].status,
-                'error'
-            )
-            assert.ok(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].content[0].json
-            )
-
-            // Verify that the history was updated correctly
-            assert.ok(secondCallArgs.conversationState?.history)
-            assert.strictEqual(secondCallArgs.conversationState?.history.length, 2)
-            assert.ok(secondCallArgs.conversationState?.history[0].userInputMessage)
-            assert.ok(secondCallArgs.conversationState?.history[1].assistantResponseMessage)
-
-            // Create expected result format matching the actual format
-            const expectedErrorChatResult: ChatResult = {
-                messageId: mockMessageId,
-                body: 'I see the tool failed with error: Tool execution failed with an error',
-                buttons: [],
-                codeReference: [],
-                header: undefined,
-                additionalMessages: [],
-            }
-
-            // Verify the final result includes both messages
-            assertChatResultsMatch(chatResult, expectedErrorChatResult)
-        })
-
-        it('handles multiple iterations of tool uses with proper history updates', async () => {
-            // First response includes a tool use request
-            const mockToolUseId1 = 'mock-tool-use-id-1'
-            const mockToolName1 = 'mock-tool-name-1'
-            const mockToolInput1 = JSON.stringify({ param1: 'value1' })
-            const mockToolResult1 = { result: 'tool execution result 1' }
-
-            // Second tool use in a subsequent response
-            const mockToolUseId2 = 'mock-tool-use-id-2'
-            const mockToolName2 = 'mock-tool-name-2'
-            const mockToolInput2 = JSON.stringify({ param2: 'value2' })
-            const mockToolResult2 = { result: 'tool execution result 2' }
-
-            // First response with first tool use
-            const mockFirstToolUseResponseList: ChatResponseStream[] = [
-                {
-                    messageMetadataEvent: {
-                        conversationId: mockConversationId,
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'I need to use tool 1. ',
-                    },
-                },
-                {
-                    toolUseEvent: {
-                        toolUseId: mockToolUseId1,
-                        name: mockToolName1,
-                        input: mockToolInput1,
-                        stop: true,
-                    },
-                },
-            ]
-
-            // Second response with second tool use
-            const mockSecondToolUseResponseList: ChatResponseStream[] = [
-                {
-                    messageMetadataEvent: {
-                        conversationId: mockConversationId,
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'Now I need to use tool 2. ',
-                    },
-                },
-                {
-                    toolUseEvent: {
-                        toolUseId: mockToolUseId2,
-                        name: mockToolName2,
-                        input: mockToolInput2,
-                        stop: true,
-                    },
-                },
-            ]
-
-            // Final response with complete answer
-            const mockFinalResponseList: ChatResponseStream[] = [
-                {
-                    messageMetadataEvent: {
-                        conversationId: mockConversationId,
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'Hello ',
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: 'World',
-                    },
-                },
-                {
-                    assistantResponseEvent: {
-                        content: '!',
-                    },
-                },
-            ]
-
-            const historyAfterTool1 = [
-                {
-                    type: 'prompt',
-                    body: 'Hello with multiple tools',
-                    userInputMessageContext: {
-                        toolResults: [],
-                    },
-                },
-                {
-                    type: 'answer',
-                    body: 'I need to use tool 1. ',
-                    toolUses: [{ toolUseId: mockToolUseId1, name: mockToolName1, input: { key: mockToolInput1 } }],
-                },
-            ]
-            const historyAfterTool2 = [
-                ...historyAfterTool1,
-                { type: 'prompt', body: 'Hello with multiple tools' },
-                {
-                    type: 'answer',
-                    body: 'Now I need to use tool 2. ',
-                    toolUses: [{ toolUseId: mockToolUseId2, name: mockToolName2, input: { key: mockToolInput2 } }],
-                },
-            ]
-
-            chatDbInitializedStub.returns(true)
-            getMessagesStub
-                .onFirstCall()
-                .returns([])
-                .onSecondCall()
-                .returns(historyAfterTool1)
-                .onThirdCall()
-                .returns(historyAfterTool2)
-
-            // Reset the stub and set up to return different responses on consecutive calls
-            generateAssistantResponseStub.restore()
-            generateAssistantResponseStub = sinon.stub(CodeWhispererStreaming.prototype, 'generateAssistantResponse')
-
-            generateAssistantResponseStub.onFirstCall().returns(
-                Promise.resolve({
-                    $metadata: {
-                        requestId: mockMessageId,
-                    },
-                    generateAssistantResponseResponse: createIterableResponse(mockFirstToolUseResponseList),
-                })
-            )
-
-            generateAssistantResponseStub.onSecondCall().returns(
-                Promise.resolve({
-                    $metadata: {
-                        requestId: mockMessageId,
-                    },
-                    generateAssistantResponseResponse: createIterableResponse(mockSecondToolUseResponseList),
-                })
-            )
-
-            generateAssistantResponseStub.onThirdCall().returns(
-                Promise.resolve({
-                    $metadata: {
-                        requestId: mockMessageId,
-                    },
-                    generateAssistantResponseResponse: createIterableResponse(mockFinalResponseList),
-                })
-            )
-
-            // Reset the runTool stub
-            const runToolStub = testFeatures.agent.runTool as sinon.SinonStub
-            runToolStub.reset()
-            runToolStub.withArgs(mockToolName1, JSON.parse(mockToolInput1)).resolves(mockToolResult1)
-            runToolStub.withArgs(mockToolName2, JSON.parse(mockToolInput2)).resolves(mockToolResult2)
-
-            // Make the request
-            const chatResultPromise = chatController.onChatPrompt(
-                { tabId: mockTabId, prompt: { prompt: 'Hello with multiple tools' } },
-                mockCancellationToken
-            )
-
-            const chatResult = await chatResultPromise
-
-            // Verify that generateAssistantResponse was called three times
-            sinon.assert.calledThrice(generateAssistantResponseStub)
-
-            // Verify that the second request included the first tool results
-            const secondCallArgs = generateAssistantResponseStub.secondCall.args[0]
-            assert.ok(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-                    .length,
-                1
-            )
-            assert.strictEqual(
-                secondCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].toolUseId,
-                mockToolUseId1
-            )
-
-            // Verify that the history was updated correctly after first tool use
-            assert.ok(secondCallArgs.conversationState?.history)
-            assert.strictEqual(secondCallArgs.conversationState?.history.length, 2)
-            assert.ok(secondCallArgs.conversationState?.history[0].userInputMessage)
-            assert.ok(secondCallArgs.conversationState?.history[1].assistantResponseMessage)
-            assert.strictEqual(
-                secondCallArgs.conversationState?.history[1].assistantResponseMessage?.content,
-                'I need to use tool 1. '
-            )
-
-            // Verify that the third request included the second tool results
-            const thirdCallArgs = generateAssistantResponseStub.thirdCall.args[0]
-            assert.ok(
-                thirdCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-            )
-            assert.strictEqual(
-                thirdCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext?.toolResults
-                    .length,
-                1
-            )
-            assert.strictEqual(
-                thirdCallArgs.conversationState?.currentMessage?.userInputMessage?.userInputMessageContext
-                    ?.toolResults[0].toolUseId,
-                mockToolUseId2
-            )
-
-            // Verify that the history was updated correctly after second tool use
-            assert.ok(thirdCallArgs.conversationState?.history)
-            assert.strictEqual(thirdCallArgs.conversationState?.history.length, 4)
-            assert.ok(thirdCallArgs.conversationState?.history[2].userInputMessage)
-            assert.ok(thirdCallArgs.conversationState?.history[3].assistantResponseMessage)
-            assert.strictEqual(
-                thirdCallArgs.conversationState?.history[3].assistantResponseMessage?.content,
-                'Now I need to use tool 2. '
-            )
-
-            // Verify the final result
-            assertChatResultsMatch(chatResult, expectedCompleteChatResult)
-        })
-
         it('returns help message if it is a help follow up action', async () => {
             const chatResultPromise = chatController.onChatPrompt(
                 { tabId: mockTabId, prompt: { prompt: DEFAULT_HELP_FOLLOW_UP_PROMPT } },
