+++ server/aws-lsp-codewhisperer/src/language-server/agenticChat/utils/promises.ts
@@ -1,0 +1,9 @@
+export const createDeferred = () => {
+    let resolve
+    let reject
+    const promise = new Promise((res, rej) => {
+        resolve = res
+        reject = (e: Error) => rej(e)
+    })
+    return { promise, resolve, reject }
+}
